Teleport players to the copy world
When the server admin didn't specifiy the lobby spawn location.
Then, when the player joins instead of the game showing a warning and that's it The game takes the player's location and then chanegs the world from original to the copy.

With this approach the game will able able to teleport players to the copy world without knowing the location exactly.